+++ src/Model/Command/Interaction.php
@@ -8,7 +8,7 @@ use Lmc\Matej\Model\Assertion;
  * Interaction command allows to send one interaction between a user and item.
  * When given user or item identifier is unknown, Matej will create such user or item respectively.
  */
-class Interaction extends AbstractCommand
+class Interaction extends AbstractCommand implements UserAwareInterface
 {
     private const INTERACTION_TYPE_DETAILVIEWS = 'detailviews';
     private const INTERACTION_TYPE_PURCHASES = 'purchases';
@@ -105,6 +105,11 @@ class Interaction extends AbstractCommand
         int $timestamp = null
     ): self {
         return new static(self::INTERACTION_TYPE_RATINGS, $userId, $itemId, $value, $context, $timestamp);
+    }
+
+    public function getUserId(): string
+    {
+        return $this->userId;
     }
 
     public function getCommandType(): string