+++ src/Model/Command/Sorting.php
@@ -8,7 +8,7 @@ use Lmc\Matej\Model\Assertion;
  * Sorting items is a way how to use Matej to deliver personalized experience to users.
  * It allows to sort given list of items according to the user preference.
  */
-class Sorting extends AbstractCommand
+class Sorting extends AbstractCommand implements UserAwareInterface
 {
     /** @var string */
     private $userId;
@@ -27,6 +27,11 @@ class Sorting extends AbstractCommand
     public static function create(string $userId, array $itemIds): self
     {
         return new static($userId, $itemIds);
+    }
+
+    public function getUserId(): string
+    {
+        return $this->userId;
     }
 
     protected function setUserId(string $userId): void